+++ ftplugin/k8s_deployments.lua
@@ -1,5 +1,5 @@
 local api = vim.api
-local actions = require("kubectl.actions.actions")
+local buffers = require("kubectl.actions.buffers")
 local commands = require("kubectl.actions.commands")
 local deployment_view = require("kubectl.views.deployments")
 local loop = require("kubectl.utils.loop")
@@ -72,7 +72,7 @@ api.nvim_buf_set_keymap(0, "n", "i", "", {
     else
       vim.ui.input({ prompt = "Update image", default = container_images[1] }, function(input)
         if input ~= nil then
-          actions.confirmation_buffer("Are you sure that you want to update the image?", nil, function(confirm)
+          buffers.confirmation_buffer("Are you sure that you want to update the image?", nil, function(confirm)
             if confirm then
               local set_image = { "set", "image", "deployment/" .. name, name .. "=" .. input, "-n", ns }
               commands.shell_command_async("kubectl", set_image, function(response)
@@ -93,7 +93,7 @@ api.nvim_buf_set_keymap(0, "n", "r", "", {
   silent = true,
   callback = function()
     local ns, name = tables.getCurrentSelection(unpack({ 1, 2 }))
-    actions.confirmation_buffer("Are you sure that you want to restart the deployment: " .. name, nil, function(confirm)
+    buffers.confirmation_buffer("Are you sure that you want to restart the deployment: " .. name, nil, function(confirm)
       if confirm then
         commands.shell_command_async("kubectl", { "rollout", "restart", "deployment/" .. name, "-n", ns }, function(response)
           vim.schedule(function()
