+++ lua/kubectl/views/deployments/init.lua
@@ -1,5 +1,5 @@
 local ResourceBuilder = require("kubectl.resourcebuilder")
-local actions = require("kubectl.actions.actions")
+local buffers = require("kubectl.actions.buffers")
 local definition = require("kubectl.views.deployments.definition")
 
 local M = {}
@@ -20,7 +20,7 @@ function M.Deployments(cancellationToken)
 end
 
 function M.Edit(name, namespace)
-  actions.floating_buffer({}, {}, "yaml", {})
+  buffers.floating_buffer({}, {}, "yaml", {})
 
   local cmd = "kubectl edit deployments/" .. name .. " -n " .. namespace
   vim.fn.termopen(cmd)
