+++ lua/kubectl/views/filter/init.lua
@@ -1,4 +1,4 @@
-local actions = require("kubectl.actions.actions")
+local buffers = require("kubectl.actions.buffers")
 local tables = require("kubectl.utils.tables")
 
 local M = {}
@@ -8,7 +8,7 @@ function M.filter()
     { key = "<enter>", desc = "apply" },
     { key = "<q>", desc = "close" },
   }, false, false)
-  actions.filter_buffer("Filter: ", marks, "k8s_filter", { title = "Filter", header = { data = header, marks = marks } })
+  buffers.filter_buffer("Filter: ", marks, "k8s_filter", { title = "Filter", header = { data = header, marks = marks } })
 end
 
 return M
