+++ lua/kubectl/views/root/init.lua
@@ -1,4 +1,4 @@
-local actions = require("kubectl.actions.actions")
+local buffers = require("kubectl.actions.buffers")
 local tables = require("kubectl.utils.tables")
 
 local M = {}
@@ -15,7 +15,7 @@ function M.Root()
   local header, marks = tables.generateHeader({
     { key = "<enter>", desc = "Select" },
   }, true, true)
-  actions.buffer(results, {}, "k8s_root", { title = "Root", header = { data = header, marks = marks } })
+  buffers.buffer(results, {}, "k8s_root", { title = "Root", header = { data = header, marks = marks } })
 end
 
 return M
