+++ lua/kubectl/views/services/init.lua
@@ -1,5 +1,5 @@
 local ResourceBuilder = require("kubectl.resourcebuilder")
-local actions = require("kubectl.actions.actions")
+local buffers = require("kubectl.actions.buffers")
 local definition = require("kubectl.views.services.definition")
 
 local M = {}
@@ -24,7 +24,7 @@ function M.Services(cancellationToken)
 end
 
 function M.Edit(name, namespace)
-  actions.floating_buffer({}, {}, "yaml", {})
+  buffers.floating_buffer({}, {}, "yaml", {})
   local cmd = "kubectl edit services/" .. name .. " -n " .. namespace
   vim.fn.termopen(cmd)
 end
